Another unsigned<<20 overflow #PSBM-43119

+++ Dispatcher/Dispatcher/Tasks/Task_EditVm_p.h
@@ -113,7 +113,7 @@ private:
 
 struct VcmmdAction: Action
 {
-	VcmmdAction(const QString& uuid_, unsigned limit_, unsigned guarantee_):
+	VcmmdAction(const QString& uuid_, quint64 limit_, quint64 guarantee_):
 		m_vcmmd(uuid_), m_limit(limit_ << 20), m_guarantee(guarantee_ << 20)
 	{
 	}
